chore: remove WalletConnectProvider and wrap web3modal state logic in ModalContext
