+++ packages/rainbowkit/src/components/RainbowKitProvider/ModalContext.tsx
@@ -13,7 +13,6 @@ import { AccountModal } from '../AccountModal/AccountModal';
 import { ChainModal } from '../ChainModal/ChainModal';
 import { ConnectModal } from '../ConnectModal/ConnectModal';
 import { useAuthenticationStatus } from './AuthenticationContext';
-import { useIsWalletConnectModalOpen } from './WalletConnectProvider';
 
 function useModalStateValue() {
   const [isModalOpen, setModalOpen] = useState(false);
@@ -32,12 +31,16 @@ interface ModalContextValue {
   openAccountModal?: () => void;
   openChainModal?: () => void;
   openConnectModal?: () => void;
+  isWalletConnectModalOpen: boolean;
+  setIsWalletConnectModalOpen: (isWalletConnectModalOpen: boolean) => void;
 }
 
 const ModalContext = createContext<ModalContextValue>({
   accountModalOpen: false,
   chainModalOpen: false,
   connectModalOpen: false,
+  isWalletConnectModalOpen: false,
+  setIsWalletConnectModalOpen: () => {},
 });
 
 interface ModalProviderProps {
@@ -62,6 +65,9 @@ export function ModalProvider({ children }: ModalProviderProps) {
     isModalOpen: chainModalOpen,
     openModal: openChainModal,
   } = useModalStateValue();
+
+  const [isWalletConnectModalOpen, setIsWalletConnectModalOpen] =
+    useState(false);
 
   const connectionStatus = useConnectionStatus();
 
@@ -105,6 +111,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
           accountModalOpen,
           chainModalOpen,
           connectModalOpen,
+          isWalletConnectModalOpen,
           openAccountModal:
             isCurrentChainSupported && connectionStatus === 'connected'
               ? openAccountModal
@@ -116,6 +123,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
             connectionStatus === 'unauthenticated'
               ? openConnectModal
               : undefined,
+          setIsWalletConnectModalOpen,
         }),
         [
           connectionStatus,
@@ -126,6 +134,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
           openChainModal,
           openConnectModal,
           isCurrentChainSupported,
+          isWalletConnectModalOpen,
         ],
       )}
     >
@@ -158,10 +167,16 @@ export function useChainModal() {
   return { chainModalOpen, openChainModal };
 }
 
+export function useWalletConnectOpenState() {
+  const { isWalletConnectModalOpen, setIsWalletConnectModalOpen } =
+    useContext(ModalContext);
+
+  return { isWalletConnectModalOpen, setIsWalletConnectModalOpen };
+}
+
 export function useConnectModal() {
   const { connectModalOpen, openConnectModal } = useContext(ModalContext);
-
-  const isWalletConnectModalOpen = useIsWalletConnectModalOpen();
+  const { isWalletConnectModalOpen } = useWalletConnectOpenState();
 
   return {
     connectModalOpen: connectModalOpen || isWalletConnectModalOpen,
