+++ packages/rainbowkit/src/components/RainbowKitProvider/RainbowKitProvider.tsx
@@ -20,7 +20,6 @@ import { RainbowKitChainProvider } from './RainbowKitChainContext';
 import { ShowBalanceProvider } from './ShowBalanceContext';
 import { ShowRecentTransactionsContext } from './ShowRecentTransactionsContext';
 import { WalletButtonProvider } from './WalletButtonContext';
-import { WalletConnectProvider } from './WalletConnectProvider';
 import { useFingerprint } from './useFingerprint';
 import { usePreloadImages } from './usePreloadImages';
 import { clearWalletConnectDeepLink } from './walletConnectDeepLink';
@@ -104,62 +103,60 @@ export function RainbowKitProvider({
 
   return (
     <RainbowKitChainProvider initialChain={initialChain}>
-      <WalletConnectProvider>
-        <WalletButtonProvider>
-          <I18nProvider locale={locale}>
-            <CoolModeContext.Provider value={coolMode}>
-              <ModalSizeProvider modalSize={modalSize}>
-                <ShowRecentTransactionsContext.Provider
-                  value={showRecentTransactions}
-                >
-                  <TransactionStoreProvider>
-                    <AvatarContext.Provider value={avatarContext}>
-                      <AppContext.Provider value={appContext}>
-                        <ThemeIdContext.Provider value={id}>
-                          <ShowBalanceProvider>
-                            <ModalProvider>
-                              {theme ? (
-                                <div {...createThemeRootProps(id)}>
-                                  <style
-                                    // biome-ignore lint/security/noDangerouslySetInnerHtml: TODO
-                                    dangerouslySetInnerHTML={{
-                                      // Selectors are sanitized to only contain alphanumeric
-                                      // and underscore characters. Theme values generated by
-                                      // cssStringFromTheme are sanitized, removing
-                                      // characters that terminate values / HTML tags.
-                                      __html: [
-                                        `${selector}{${cssStringFromTheme(
-                                          'lightMode' in theme
-                                            ? theme.lightMode
-                                            : theme,
-                                        )}}`,
+      <WalletButtonProvider>
+        <I18nProvider locale={locale}>
+          <CoolModeContext.Provider value={coolMode}>
+            <ModalSizeProvider modalSize={modalSize}>
+              <ShowRecentTransactionsContext.Provider
+                value={showRecentTransactions}
+              >
+                <TransactionStoreProvider>
+                  <AvatarContext.Provider value={avatarContext}>
+                    <AppContext.Provider value={appContext}>
+                      <ThemeIdContext.Provider value={id}>
+                        <ShowBalanceProvider>
+                          <ModalProvider>
+                            {theme ? (
+                              <div {...createThemeRootProps(id)}>
+                                <style
+                                  // biome-ignore lint/security/noDangerouslySetInnerHtml: TODO
+                                  dangerouslySetInnerHTML={{
+                                    // Selectors are sanitized to only contain alphanumeric
+                                    // and underscore characters. Theme values generated by
+                                    // cssStringFromTheme are sanitized, removing
+                                    // characters that terminate values / HTML tags.
+                                    __html: [
+                                      `${selector}{${cssStringFromTheme(
+                                        'lightMode' in theme
+                                          ? theme.lightMode
+                                          : theme,
+                                      )}}`,
 
-                                        'darkMode' in theme
-                                          ? `@media(prefers-color-scheme:dark){${selector}{${cssStringFromTheme(
-                                              theme.darkMode,
-                                              { extends: theme.lightMode },
-                                            )}}}`
-                                          : null,
-                                      ].join(''),
-                                    }}
-                                  />
-                                  {children}
-                                </div>
-                              ) : (
-                                children
-                              )}
-                            </ModalProvider>
-                          </ShowBalanceProvider>
-                        </ThemeIdContext.Provider>
-                      </AppContext.Provider>
-                    </AvatarContext.Provider>
-                  </TransactionStoreProvider>
-                </ShowRecentTransactionsContext.Provider>
-              </ModalSizeProvider>
-            </CoolModeContext.Provider>
-          </I18nProvider>
-        </WalletButtonProvider>
-      </WalletConnectProvider>
+                                      'darkMode' in theme
+                                        ? `@media(prefers-color-scheme:dark){${selector}{${cssStringFromTheme(
+                                            theme.darkMode,
+                                            { extends: theme.lightMode },
+                                          )}}}`
+                                        : null,
+                                    ].join(''),
+                                  }}
+                                />
+                                {children}
+                              </div>
+                            ) : (
+                              children
+                            )}
+                          </ModalProvider>
+                        </ShowBalanceProvider>
+                      </ThemeIdContext.Provider>
+                    </AppContext.Provider>
+                  </AvatarContext.Provider>
+                </TransactionStoreProvider>
+              </ShowRecentTransactionsContext.Provider>
+            </ModalSizeProvider>
+          </CoolModeContext.Provider>
+        </I18nProvider>
+      </WalletButtonProvider>
     </RainbowKitChainProvider>
   );
 }
