+++ packages/rainbowkit/src/wallets/useWalletConnectors.ts
@@ -1,6 +1,6 @@
 import { Config, Connector, useConnect } from 'wagmi';
 import { ConnectMutateAsync } from 'wagmi/query';
-import { useWalletConnectOpenState } from '../components/RainbowKitProvider/WalletConnectProvider';
+import { useWalletConnectOpenState } from '../components/RainbowKitProvider/ModalContext';
 import { indexBy } from '../utils/indexBy';
 import {
   useInitialChainId,
